requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
httpx
redis[async]
celery[redis]
openai
alembic
python-multipart
aiofiles
python-jose[cryptography]
passlib[bcrypt]
SpeechRecognition
gTTS
pydub
psycopg2-binary
